place new helper method in context of report view helper
remove one unused method and take all report oriented methods into report view helper, while leaving context switching in the grab bag helper

goal of this helper is to NOT include all report specific computations--those should go in components/methods of report classes. this is just to manage formatting view concerns.
