--- app/app/helpers/money_helper.rb
@@ -1,6 +1,0 @@
-module MoneyHelper
-  def format_hours(hour)
-    return hour unless Float(hour, exception: false).present?
-    hour.to_f.round(1)
-  end
-end
