+++ app/app/helpers/report_view_helper.rb
@@ -1,0 +1,49 @@
+module ReportViewHelper
+  def format_hours(hour)
+    return hour unless Float(hour, exception: false).present?
+    hour.to_f.round(1)
+  end
+
+  def format_parsed_date(date, format = :default)
+    return unless date
+    I18n.l(date.to_date, format: format)
+  end
+
+  def format_date(timestamp_string, format = :long)
+    begin
+      parsed_time = timestamp_string.is_a?(String) ? Time.parse(timestamp_string) : timestamp_string
+      I18n.l(parsed_time.to_date, format: format)
+    rescue
+      timestamp_string
+    end
+  end
+
+  def format_money(dollars_in_cents)
+    number_to_currency(dollars_in_cents.to_f / 100)
+  end
+
+  def translate_aggregator_value(namespace, value)
+    return unless value.present?
+
+    i18n_key = "aggregator_strings.#{namespace}.#{value}"
+
+    # convert the key to snake_case, replacing hyphens with underscores
+    i18n_key = i18n_key.gsub("-", "_").downcase
+
+    if I18n.exists?(i18n_key)
+      I18n.t(i18n_key)
+    elsif I18n.locale == :en
+      # if the key isn't in spanish, just return the original value
+      value
+    else
+      if Rails.env.development? || Rails.env.test?
+        raise "Missing aggregator translation for #{namespace}.#{value}"
+      end
+
+      # In production, log warning and return original value
+      Rails.logger.warn "Unknown aggregator value for #{namespace}: #{value}"
+
+      value
+    end
+  end
+end
